fixed some ambiguous method calls passing a uint64_t  for a long

+++ iod/src/IOComponent.cpp
@@ -428,8 +428,8 @@ int32_t AnalogueInput::filter(int32_t raw) {
 	std::list<MachineInstance*>::iterator owners_iter = owners.begin();
 	while (owners_iter != owners.end()) {
 		MachineInstance *o = *owners_iter++;
-		o->properties.add("IOTIME", read_time, SymbolTable::ST_REPLACE);
-		o->properties.add("IOVALUE", raw, SymbolTable::ST_REPLACE);
+		o->properties.add("IOTIME", (long)read_time, SymbolTable::ST_REPLACE);
+		o->properties.add("IOVALUE", (long)raw, SymbolTable::ST_REPLACE);
 	}
     config->positions.append(raw);
     int32_t mean = (config->positions.average(config->buffer_len) + 0.5f);
@@ -456,8 +456,8 @@ int32_t Counter::filter(int32_t val) {
 	std::list<MachineInstance*>::iterator owners_iter = owners.begin();
 	while (owners_iter != owners.end()) {
 		MachineInstance *o = *owners_iter++;
-		o->properties.add("IOTIME", read_time, SymbolTable::ST_REPLACE);
-		o->properties.add("IOVALUE", val, SymbolTable::ST_REPLACE);
+		o->properties.add("IOTIME", (long)read_time, SymbolTable::ST_REPLACE);
+		o->properties.add("IOVALUE", (long)val, SymbolTable::ST_REPLACE);
 	}
 	return IOComponent::filter(val);
 }
@@ -475,8 +475,8 @@ int32_t CounterRate::filter(int32_t val) {
 	std::list<MachineInstance*>::iterator owners_iter = owners.begin();
 	while (owners_iter != owners.end()) {
 		MachineInstance *o = *owners_iter++;
-		o->properties.add("IOTIME", read_time, SymbolTable::ST_REPLACE);
-		o->properties.add("IOVALUE", val, SymbolTable::ST_REPLACE);
+		o->properties.add("IOTIME", (long)read_time, SymbolTable::ST_REPLACE);
+		o->properties.add("IOVALUE", (long)val, SymbolTable::ST_REPLACE);
 	}
     return IOComponent::filter(val);
 /* disabled since this is now implemented at the MachineInstance level
